Code Analyzer
Code Analyzer is the C#(.net) Application to analyze other C#
applications in terms of size and complexity. This managed code analyzer
finds all the types, their members, and relationships with other types.
It constructs two metrics for each member function: its size in source
lines of code (SLOC), and cyclomatic complexity defined as the number of
elements in its scope tree.
The purpose of this code analysis is to find functions that may need
attention because of their size and/or complexity.
Code Analyzer is a powerful tool to understand the structure and design
of the code as it reveals the relationship and interactions between all
the types. Thus giving a reference point for various groups of users
such developers, quality assurance testers and managers. It provides the
customizable GUI for different set of users to pull out the relative
information about the code according to the need and responsibility of
different users.